Extract browser surface state
