+++ Sources/QuillCodeApp/QuillCodeBrowserSurface.swift
@@ -1,0 +1,89 @@
+import Foundation
+import QuillCodeCore
+
+public struct BrowserCommentState: Sendable, Hashable, Identifiable {
+    public var id: UUID
+    public var url: String
+    public var text: String
+    public var createdAt: Date
+
+    public init(id: UUID = UUID(), url: String, text: String, createdAt: Date = Date()) {
+        self.id = id
+        self.url = url
+        self.text = text
+        self.createdAt = createdAt
+    }
+}
+
+public struct BrowserSnapshotState: Sendable, Hashable {
+    public var sourceLabel: String
+    public var inspectionDepth: BrowserInspectionDepth
+    public var summary: String
+    public var details: [String]
+    public var outline: [String]
+    public var textSnippet: String?
+
+    public init(
+        sourceLabel: String,
+        inspectionDepth: BrowserInspectionDepth = .metadataOnly,
+        summary: String,
+        details: [String] = [],
+        outline: [String] = [],
+        textSnippet: String? = nil
+    ) {
+        self.sourceLabel = sourceLabel
+        self.inspectionDepth = inspectionDepth
+        self.summary = summary
+        self.details = details
+        self.outline = outline
+        self.textSnippet = textSnippet
+    }
+}
+
+public struct BrowserState: Sendable, Hashable {
+    public var isVisible: Bool
+    public var addressDraft: String
+    public var currentURL: String?
+    public var history: [String]
+    public var historyIndex: Int?
+    public var title: String
+    public var status: String
+    public var snapshot: BrowserSnapshotState?
+    public var comments: [BrowserCommentState]
+
+    public var canGoBack: Bool {
+        guard let historyIndex else { return false }
+        return history.indices.contains(historyIndex) && historyIndex > history.startIndex
+    }
+
+    public var canGoForward: Bool {
+        guard let historyIndex else { return false }
+        return history.indices.contains(historyIndex) && history.index(after: historyIndex) < history.endIndex
+    }
+
+    public var canReload: Bool {
+        currentURL != nil
+    }
+
+    public init(
+        isVisible: Bool = false,
+        addressDraft: String = "",
+        currentURL: String? = nil,
+        history: [String] = [],
+        historyIndex: Int? = nil,
+        title: String = "Browser preview",
+        status: String = "Ready",
+        snapshot: BrowserSnapshotState? = nil,
+        comments: [BrowserCommentState] = []
+    ) {
+        self.isVisible = isVisible
+        self.addressDraft = addressDraft
+        self.currentURL = currentURL
+        self.history = history
+        self.historyIndex = historyIndex
+        self.title = title
+        self.status = status
+        self.snapshot = snapshot
+        self.comments = comments
+    }
+}
