+++ Tests/QuillCodeParityTests/ParityGateTests.swift
@@ -41,6 +41,19 @@ final class ParityGateTests: XCTestCase {
         XCTAssertFalse(modelText.contains("public struct ToolArtifactState"), "WorkspaceModel should not own tool artifact surface state.")
     }
 
+    func testWorkspaceModelDelegatesBrowserSurfaceTypes() throws {
+        let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
+        let browserSurfaceText = try Self.appSourceText(named: "QuillCodeBrowserSurface.swift")
+
+        XCTAssertTrue(browserSurfaceText.contains("public struct BrowserState"), "Browser state should live in a focused surface file.")
+        XCTAssertTrue(browserSurfaceText.contains("public struct BrowserSnapshotState"), "Browser snapshot state should live in a focused surface file.")
+        XCTAssertTrue(browserSurfaceText.contains("public struct BrowserCommentState"), "Browser comment state should live in a focused surface file.")
+        XCTAssertTrue(modelText.contains("BrowserInspector.snapshot"), "WorkspaceModel should still orchestrate browser snapshot refreshes.")
+        XCTAssertFalse(modelText.contains("public struct BrowserState"), "WorkspaceModel should not own browser surface state.")
+        XCTAssertFalse(modelText.contains("public struct BrowserSnapshotState"), "WorkspaceModel should not own browser snapshot state.")
+        XCTAssertFalse(modelText.contains("public struct BrowserCommentState"), "WorkspaceModel should not own browser comment state.")
+    }
+
     func testDesktopDefinesNativeMenuBarWidget() throws {
         let text = try Self.desktopSourceText()
